requirements.txt
gradio==3.32.0
Pillow==9.5.0
pytesseract==0.3.10
